requirements.txt
python-dotenv>=1.0.1
openai>=1.85.0
gradio>=5.33.1
pypdf>=4.3.1
faiss-cpu==1.8.0.post1
numpy>=1.26.4
requests>=2.32.3
fastapi>=0.115
uvicorn>=0.30
nbformat>=5.10.4
pdf2image>=1.17.0
pillow
